% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{customerByDate_query}
\alias{customerByDate_query}
\title{按日期查询}
\usage{
customerByDate_query(token, FDate)
}
\arguments{
\item{token}{中台数据库token}

\item{FDate}{日期}
}
\value{

}
\description{
按日期查询
}
\examples{
customerByDate_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,2023-07-03)
}
