% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerByDatequery}
\alias{customerByDatequery}
\title{客户按日期查询}
\usage{
customerByDatequery(token, fdate)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fdate}{日期}
}
\value{
返回值
}
\description{
客户按日期查询
}
\examples{
customerByDatequery('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','2023-07-11')

}
