% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerByDatesync}
\alias{customerByDatesync}
\title{客户按日期同步}
\usage{
customerByDatesync(token, fdate)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fdate}{日期}
}
\value{
返回值
}
\description{
客户按日期同步
}
\examples{
customerByDatesync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','2023-07-11')

}
