% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerByFNamesync}
\alias{customerByFNamesync}
\title{'客户按名称同步}
\usage{
customerByFNamesync(token, FCustomerName)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fdate}{单据编号}
}
\value{
返回值
}
\description{
'客户按名称同步
}
\examples{
customerByFNamesync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','宁夏谱合科技有限公司')

}
