% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{customerByFNumber_query}
\alias{customerByFNumber_query}
\title{按单据编号查询}
\usage{
customerByFNumber_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
按单据编号查询
}
\examples{
customerByFNumber_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,C15276)
}
