% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerByFNumberquery}
\alias{customerByFNumberquery}
\title{客户按单据编号查询}
\usage{
customerByFNumberquery(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
客户按单据编号查询
}
\examples{
materialByFNumberquery('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C15276')

}
