% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{customerErpDataByFNumber_query}
\alias{customerErpDataByFNumber_query}
\title{按编号查询}
\usage{
customerErpDataByFNumber_query(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按编号查询
}
\examples{
customerErpDataByFNumber_query(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,C000186,赛普集团新账套)
}
