% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerErpDataByFNumberquery}
\alias{customerErpDataByFNumberquery}
\title{客户erp数据查询}
\usage{
customerErpDataByFNumberquery(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
客户erp数据查询
}
\examples{
customerErpDataByFNumberquery('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C009170')
}
