% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/customer.R
\name{customerLog_query}
\alias{customerLog_query}
\title{日志同步}
\usage{
customerLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志同步
}
\examples{
customerLog_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C15276')
}
