% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{customerStatusupload}
\alias{customerStatusupload}
\title{客户修改单据状态}
\usage{
customerStatusupload(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
客户修改单据状态
}
\examples{
mdstatusuploadbyno('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C15276')

}
