% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialByDate_sync}
\alias{materialByDate_sync}
\title{按日期同步}
\usage{
materialByDate_sync(token, FDate, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FDate}{日期}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按日期同步
}
\examples{
materialByDate_sync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','2022-12-01','赛普集团新账套')
}
