% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialByFNumber_sync}
\alias{materialByFNumber_sync}
\title{按单号同步}
\usage{
materialByFNumber_sync(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单号同步
}
\examples{
materialByFNumber_sync(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,K86-10015S-H,赛普集团新账套)
}
