% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{materialByFNumberSync}
\alias{materialByFNumberSync}
\title{'物料按编号同步}
\usage{
materialByFNumberSync(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fdate}{单据编号}
}
\value{
返回值
}
\description{
'物料按编号同步
}
\examples{
materialByFNumbersync('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','NZ0101')

}
