% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialLog_query}
\alias{materialLog_query}
\title{日志同步}
\usage{
materialLog_query(token, FNumber)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}
}
\value{

}
\description{
日志同步
}
\examples{
materialLog_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','K86-10015S-H')
}
