% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialStatus_upload}
\alias{materialStatus_upload}
\title{修改数据状态}
\usage{
materialStatus_upload(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
修改数据状态
}
\examples{
materialStatus_upload('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C15276','赛普集团新账套')
}
