% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{md_erpdataquery_byno}
\alias{md_erpdataquery_byno}
\title{erp数据查询}
\usage{
md_erpdataquery_byno(token, category, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{category}{单据类型(客户，物料，供应商)}

\item{fnumber}{单据编号}
}
\value{

}
\description{
erp数据查询
}
