% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{mdsrcquerybydate}
\alias{mdsrcquerybydate}
\title{按日期查询}
\usage{
mdsrcquerybydate(token, category, fdate)
}
\arguments{
\item{token}{中台数据库密钥}

\item{category}{单据类型(客户，物料，供应商)}

\item{fdate}{日期}
}
\value{
返回值
}
\description{
按日期查询
}
\examples{
mdsrcquerybydate('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','物料','2023-07-11')

}
