% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{mdsrcquerybyno}
\alias{mdsrcquerybyno}
\title{按单据编号查询}
\usage{
mdsrcquerybyno(token, category, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{category}{单据类型(客户，物料，供应商)}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
按单据编号查询
}
\examples{
mdsrcquerybyno('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','物料','NZ0101')

}
