% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{mdsyncbyno}
\alias{mdsyncbyno}
\title{'按编号同步}
\usage{
mdsyncbyno(token, category, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{category}{单据类型(客户，物料，供应商)}

\item{fdate}{单据编号}
}
\value{
返回值
}
\description{
'按编号同步
}
\examples{
mdsyncbyno('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','物料','NZ0101')

}
