% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplierByDate_query}
\alias{supplierByDate_query}
\title{按日期查询}
\usage{
supplierByDate_query(token, FDate)
}
\arguments{
\item{token}{中台数据库token}

\item{FDate}{日期}
}
\value{

}
\description{
按日期查询
}
\examples{
supplierByDate_query('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','2023-07-03')
}
