% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplierByDate_sync}
\alias{supplierByDate_sync}
\title{按日期同步}
\usage{
supplierByDate_sync(token, FDate, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FDate}{日期}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按日期同步
}
\examples{
supplierByDate_sync(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,2023-02-06,赛普集团新账套)
}
