% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{supplierByFNamesync}
\alias{supplierByFNamesync}
\title{'供应商按名称同步}
\usage{
supplierByFNamesync(token, FSupplierName)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fdate}{单据编号}
}
\value{
返回值
}
\description{
'供应商按名称同步
}
\examples{
mdsyncbyno('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','河间市龙腾塑料制品有限公司')

}
