% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{supplierByFNumberquery}
\alias{supplierByFNumberquery}
\title{供应商按单据编号查询}
\usage{
supplierByFNumberquery(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
供应商按单据编号查询
}
\examples{
supplierByFNumberquery('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','C15276')

}
