% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecsmd.R
\name{supplierErpDataByFNumberquery}
\alias{supplierErpDataByFNumberquery}
\title{供应商erp数据查询}
\usage{
supplierErpDataByFNumberquery(token, fnumber)
}
\arguments{
\item{token}{中台数据库密钥}

\item{fnumber}{单据编号}
}
\value{
返回值
}
\description{
供应商erp数据查询
}
\examples{
supplierErpDataByFNumberquery('9B6F803F-9D37-41A2-BDA0-70A7179AF0F3','V001851')
}
