% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplierStatus_upload}
\alias{supplierStatus_upload}
\title{修改数据状态}
\usage{
supplierStatus_upload(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
修改数据状态
}
