% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/material.R
\name{materialByFNumber_sync}
\alias{materialByFNumber_sync}
\title{按单号同步}
\usage{
materialByFNumber_sync(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单号同步
}
