% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/supplier.R
\name{supplierByFNumber_sync}
\alias{supplierByFNumber_sync}
\title{按单号同步}
\usage{
supplierByFNumber_sync(token, FNumber, FName)
}
\arguments{
\item{token}{中台数据库token}

\item{FNumber}{单据编号}

\item{FName}{ERP账套名}
}
\value{

}
\description{
按单号同步
}
\examples{
supplierByFNumber_sync(9B6F803F-9D37-41A2-BDA0-70A7179AF0F3,S00085,赛普集团新账套)
}
